Renamed classes and tests
